[QCF-102] JOB_LIST util로 이동
